feat!: upgrade to prettier v3 (#516)
