+++ .changeset/orange-eyes-brush.md
@@ -1,0 +1,5 @@
+---
+"eslint-plugin-prettier": major
+---
+
+feat!: upgrade to prettier v3
